requirements.txt
requests
Pillow
pyquery